third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-8.html

<html>
<head>
<script src="../../inspector/inspector-test.js"></script>
<script src="../../inspector/debugger-test.js"></script>
<script src="../../inspector/sources-test.js"></script>

<script>

function test() {
  var testJSFormatter = SourcesTestRunner.testPrettyPrint.bind(InspectorTest, 'text/javascript');

  TestRunner.runTestSuite([
    function asyncAwaitSupport(next) {
      var mappingQueries = ['async', 'function', 'foo', 'return', 'Promise', 'resolve'];
      testJSFormatter('async function foo() {return await Promise.resolve(1);}', mappingQueries, next);
    },
  ]);
}

</script>

</head>

<body onload="runTest()">
<p>Verifies JavaScript pretty-printing functionality.</p>
</body>
</html>
