third_party/WebKit/LayoutTests/http/tests/devtools/sources/pretty-print-javascript-8-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies JavaScript pretty-printing functionality.


Running: asyncAwaitSupport
====== 8< ------
async function foo() {
    return await Promise.resolve(1);
}

------ >8 ======
Correct mapping for <async>
Correct mapping for <function>
Correct mapping for <foo>
Correct mapping for <return>
Correct mapping for <Promise>
Correct mapping for <resolve>
